refactor: move webhook payload types under src/webhooks/github/
